feat: LectureDetail의 자식 컴포넌트 디렉터리 위치 이동 및 헬퍼 함수 정리
1. LectureDetail의 하위 구성요소(LectureHero, LectureContent, LectureCurriculum, LectureTabs)를
   `components/lecture/` 디렉터리로 이동하여 구조 일관성 확보
2. 반복적으로 사용되는 헬퍼 함수를 `utils/` 내로 묶어 관리 (getCategoryName 등)
