+++ src/components/lecture/LectureContent.jsx
