+++ src/components/lecture/LectureCurriculum.jsx
