+++ src/components/lecture/LectureTabs.jsx
